Fix 130+ compilation warnings: memcpy/memset, sign-compare, and unused variables
- Replace memcpy/memset with std::copy_n/std::fill_n for Complex types (class-memaccess warnings)
- Fix sign-compare warnings by changing int loop variables to size_t and adding explicit casts
- Remove unused NumThreads variables from BBFGPermanentCalculator*
- Comment out unused static timing variables in CGeneralizedClifford* strategies
- Fix static function declarations in GaussianSimulationStrategy and ThresholdBosonSampling headers

Files modified: 22 files
Warnings eliminated: 130+

+++ piquassoboost/sampling/simulation_strategies/source/ThresholdBosonSampling.cpp
@@ -593,7 +593,7 @@ ThresholdBosonSampling::calc_Qinv( matrix_real& matrix, double& Qdet ) {
     //  calculate the determinant of Q
     Qdet = 1.0;
     for (size_t idx=0; idx<Q.rows; idx++) {
-        if (ipiv[idx] != idx+1) {
+        if (ipiv[idx] != (int)(idx+1)) {
             Qdet = -Qdet * Q[idx*Q.stride + idx];
         }
         else {
